define GuiElement interface
this can be used to distinguish between things that have a real world component (Devices)
and things that are purely conceptual to control other aspects of the system (GuiElement)

All devices should be able to expose something about themselves as html, either by offering
buttond to control their state, or, in the case of sensors, by describing what state they
are currently in.

In this model, physical inputs such as remotes, that only give commands
are not countes as devices.