Create a module structure to support multiple complex environments
This allows for arbitrarily complex device groups, while only instatiating and displaying
devices that are present in the configured home.

The implemented solution is only viable for small numbers of environments, and would
likely require a stronger configuration interface if it were to be deployed across
more households

which it will not.